Move SRI test into separate suite
