+++ test/e2e/app-dir/app/next.config.js
@@ -1,9 +1,6 @@
 module.exports = {
   experimental: {
     clientRouterFilterRedirects: true,
-    sri: {
-      algorithm: 'sha256',
-    },
     webpackBuildWorker: true,
   },
   // output: 'standalone',
