+++ test/production/app-dir/subresource-integrity/fixture/app/dashboard/client-comp.module.css
@@ -1,0 +1,3 @@
+.client {
+  color: #2e5aea;
+}
