+++ test/production/app-dir/subresource-integrity/fixture/app/dashboard/global.css
@@ -1,0 +1,3 @@
+.dangerous-text {
+  color: red;
+}
