+++ test/production/app-dir/subresource-integrity/fixture/app/dashboard/style.css
@@ -1,0 +1,3 @@
+.green {
+  color: green;
+}
