+++ test/production/app-dir/subresource-integrity/fixture/app/page.tsx
@@ -1,0 +1,3 @@
+export default function Page() {
+  return <p>hello world</p>
+}
